+++ src/cjs/utils/datastructures/Girih.js
@@ -6,6 +6,7 @@
  * @modified 2024-03-10 Fixed some types for Typescript 5 compatibility.
  * @modified 2026-01-06 Added method `Girih.locateContainingTileAndEdge` to locate tile/edge pairs.
  * @modified 2026-01-12 Added method `Girih.getTileByCenter` to locate tiles by position.
+ * @modified 2026-01-18 Added method `Girih.removeAllTiles`.
  * @version  1.1.0
  * @file     Girih
  **/
@@ -97,6 +98,17 @@ var Girih = /** @class */ (function () {
      */
     Girih.prototype.removeTileAt = function (index) {
         this.tiles.splice(index, 1);
+    };
+    /**
+     * Remove all tiles.
+     *
+     * @name removeAllTiles
+     * @memberof Girih
+     * @instance
+     * @return {void}
+     */
+    Girih.prototype.removeAllTiles = function () {
+        this.tiles.splice(0, this.tiles.length);
     };
     /**
      * Replace all current tiles with the given ones.